requirements.txt
streamlit==1.29.0
pandas==2.0.3
scikit-learn==1.1.2
matplotlib==3.7.4
seaborn==0.13.0
joblib==1.1.1
pymc3==3.11.5
numpy==1.22.1
theano-pymc==1.1.2